+++ tests/unit/entrypoint/database/test_download.py
@@ -1,10 +1,30 @@
+# Blue Brain Search is a text mining toolbox focused on scientific use cases.
+#
+# Copyright (C) 2020  Blue Brain Project, EPFL.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program. If not, see <https://www.gnu.org/licenses/>.
 import argparse
+import datetime
 import inspect
+import logging
 import pathlib
-from datetime import datetime
+from collections import namedtuple
+from itertools import chain
 from unittest.mock import Mock
 
 import pytest
+from google.cloud.storage import Blob
 
 from bluesearch.entrypoint.database import download
 
@@ -19,7 +39,7 @@ def test_init_parser():
 
     # Test the values
     assert args.source == "pmc"
-    assert args.from_month == datetime(2020, 10, 1)
+    assert args.from_month == datetime.datetime(2020, 10, 1)
     assert args.output_dir == pathlib.Path("/path/to/download")
     assert not args.dry_run
 
@@ -51,7 +71,7 @@ def test_pmc_download(capsys, monkeypatch, tmp_path):
         "bluesearch.database.download.generate_pmc_urls", fake_generate_pmc_urls
     )
 
-    fake_datetime = datetime(2021, 11, 1)
+    fake_datetime = datetime.datetime(2021, 12, 2)
     pmc_path = tmp_path / "pmc"
     download.run("pmc", fake_datetime, pmc_path, dry_run=False)
     assert pmc_path.exists()
@@ -93,7 +113,7 @@ def test_pubmed_download(capsys, monkeypatch, tmp_path):
         "bluesearch.database.download.get_pubmed_urls", fake_get_pubmed_urls
     )
 
-    fake_datetime = datetime(2021, 11, 1)
+    fake_datetime = datetime.datetime(2021, 12, 16)
     pubmed_path = tmp_path / "pubmed"
 
     # Run the command
@@ -130,7 +150,7 @@ def test_biorxiv_medrxiv_download(source, monkeypatch, tmp_path, capsys):
     # Define mocks
     fake_getpass = Mock()
     fake_getpass.getpass.return_value = "somecredentials"
-    fake_datetime = datetime(2021, 11, 1)
+    fake_datetime = datetime.datetime(2021, 11, 1)
 
     fake_get_s3_urls = Mock(
         return_value={
@@ -180,3 +200,88 @@ def test_biorxiv_medrxiv_download(source, monkeypatch, tmp_path, capsys):
 
     stdout = capsys.readouterr().out
     assert "Month: November_2018\n1.meca\nMonth: December_2018\n2.meca\n" in stdout
+
+
+class TestArxivDownload:
+    @pytest.fixture
+    def mocked(self, monkeypatch):
+        """Set up mocks for all subsequent tests.
+
+        Need to mock:
+        * The Google Cloud Storage client and bucket
+        * get_gcs_urls
+        * download_gcs_blob
+
+        For a more convenient access pack all mocks into a namedtuple and yield.
+        """
+        client = Mock()
+        bucket = Mock()
+        get_gcs_urls = Mock()
+        get_gcs_urls.return_value = {
+            "2111": [Blob("blob_{i}", bucket) for i in range(101)],
+            "2112": [Blob("blob_x1", bucket), Blob("blob_x2", bucket)],
+        }
+        download_gcs_blob = Mock()
+        monkeypatch.setattr("google.cloud.storage.Client", client)
+        monkeypatch.setattr(
+            "bluesearch.database.download.download_gcs_blob",
+            download_gcs_blob,
+        )
+        monkeypatch.setattr(
+            "bluesearch.database.download.get_gcs_urls",
+            get_gcs_urls,
+        )
+
+        Mocked = namedtuple("Mocked", ["client", "get_gcs_urls", "download_gcs_blob"])
+        yield Mocked(client, get_gcs_urls, download_gcs_blob)
+
+    def test_dry_run(self, capsys, tmp_path, mocked):
+        # The dry run should print all months and blobs to stdout
+        download.run("arxiv", datetime.datetime(2021, 12, 1), tmp_path, dry_run=True)
+        stdout, stderr = capsys.readouterr()
+        for month, blobs in mocked.get_gcs_urls().items():
+            assert month in stdout
+            for blob in blobs:
+                assert blob.name in stdout
+
+    def test_normal_run(self, tmp_path, mocked):
+        # Under normal circumstances download_gcs_blob should be called as
+        # many times as there are blobs to download
+        download.run("arxiv", datetime.datetime(2021, 12, 1), tmp_path, dry_run=False)
+        n_blobs = sum(len(blobs) for blobs in mocked.get_gcs_urls().values())
+        assert mocked.download_gcs_blob.call_count == n_blobs
+
+    def test_worker_errors_are_reported(self, caplog, tmp_path, mocked):
+        # If a worker (download_gcs_blob) raises an error then it should
+        # appear in the error logs
+        error_msg = "something went wrong"
+        mocked.download_gcs_blob.side_effect = RuntimeError(error_msg)
+        with caplog.at_level(logging.ERROR):
+            download.run(
+                "arxiv", datetime.datetime(2021, 12, 1), tmp_path, dry_run=False
+            )
+        assert error_msg in caplog.text
+        for blob in chain(*mocked.get_gcs_urls().values()):
+            assert f"{blob.name} failed" in caplog.text
+
+
+@pytest.mark.parametrize(
+    ("source", "expected_date"),
+    [
+        ("arxiv", "April 2007"),
+        ("biorxiv", "December 2018"),
+        ("medrxiv", "October 2020"),
+        ("pmc", "December 2021"),
+        ("pubmed", "December 2021"),
+    ],
+)
+def test_structure_change(source, expected_date, tmp_path, caplog):
+
+    limit_datetime = download.MIN_DATE[source]
+    fake_datetime = limit_datetime - datetime.timedelta(days=32)
+
+    with caplog.at_level(logging.ERROR):
+        exit_code = download.run(source, fake_datetime, tmp_path, dry_run=False)
+
+    assert exit_code == 1
+    assert expected_date in caplog.text