+++ tests/unit/test_utils.py
@@ -16,9 +16,11 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
+from __future__ import annotations
 
 import json
 import pathlib
+from typing import Any
 
 import h5py
 import numpy as np
@@ -30,9 +32,40 @@ from bluesearch.utils import (
     JSONL,
     Timer,
     check_entity_type_consistency,
+    find_files,
     get_available_spacy_models,
     load_spacy_model,
 )
+
+
+class TestFindFiles:
+    def test_filtering_recursive(tmp_path):
+        input_path = pathlib.Path("tests/data/cord19_v35/document_parses/pdf_json/")
+        inputs = find_files(input_path, recursive=True)
+        filenames = sorted(x.name for x in inputs)
+        expected = [
+            "16e82ce0e0c8a1b36497afc0d4392b4fe21eb174.json",
+            "5f267fa1ef3a65e239aa974329e935a4d93dafd2.json",
+        ]
+        assert filenames == expected
+
+    def test_filtering(tmp_path):
+        input_path = pathlib.Path("tests/data/cord19_v35/")
+        inputs = find_files(
+            input_path, recursive=True, match_filename=r"[a-z0-9]+\.json"
+        )
+        filenames = sorted(x.name for x in inputs)
+        expected = [
+            "16e82ce0e0c8a1b36497afc0d4392b4fe21eb174.json",
+            "5f267fa1ef3a65e239aa974329e935a4d93dafd2.json",
+        ]
+        assert filenames == expected
+
+    def test_filtering_empty(tmp_path):
+        message = "Value for argument 'match-filename' should not be empty!"
+        input_path = pathlib.Path("tests/data/cord19_v35/")
+        with pytest.raises(ValueError, match=message):
+            _ = find_files(input_path, recursive=True, match_filename="")
 
 
 class TestTimer:
@@ -345,7 +378,7 @@ class TestH5:
 def test_load_save_jsonl(tmpdir):
     path = pathlib.Path(str(tmpdir)) / "file.jsonl"
 
-    li = [{"a": 1, "b": "cc"}, {"k": 23}]
+    li: list[dict[str, Any]] = [{"a": 1, "b": "cc"}, {"k": 23}]
     JSONL.dump_jsonl(li, path)
     lo = JSONL.load_jsonl(path)